+++ src/App/Bootstrap.php
@@ -2,9 +2,11 @@
 
 namespace App;
 
+use Dotenv\Dotenv;
 use Framework\Config;
 use Framework\Logger;
 use MiaKiwi\Kaphpir\Settings\DefaultSettings;
+use Pecee\SimpleRouter\SimpleRouter;
 
 
 
@@ -13,7 +15,41 @@ if (!isset($_ENV['LOG_LEVEL']) || $_ENV['LOG_LEVEL'] !== 'debug') {
     // Hide all errors and warnings
     error_reporting(0);
     ini_set('display_errors', '0');
+} else {
+    // Show all errors and warnings
+    error_reporting(E_ALL);
+    ini_set('display_errors', '1');
 }
+
+set_error_handler(function ($errno, $errstr, $errfile, $errline) {
+    // Log the error
+    Logger::get()->error("Error occurred", [
+        'errno' => $errno,
+        'errstr' => $errstr,
+        'errfile' => $errfile,
+        'errline' => $errline
+    ]);
+});
+
+
+
+// Load the Composer autoloader
+require_once __DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'vendor' . DIRECTORY_SEPARATOR . 'autoload.php';
+
+
+
+// Load the environment variables
+$dotenv = Dotenv::createImmutable(__DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . '..');
+$dotenv->load();
+
+
+
+// Load the application configuration
+Logger::get()->debug("Loading application configuration from file", [
+    'file' => $_ENV['APP_CONFIG']
+]);
+
+Config::load($_ENV['APP_CONFIG']);
 
 
 
@@ -31,4 +67,15 @@ foreach ($kaphpir_settings as $key => $value) {
 
 
 
+// Import the routes
 require_once __DIR__ . DIRECTORY_SEPARATOR . 'Routes.php';
+
+
+
+// Start the router
+Logger::get()->debug("---------- Received request ----------", [
+    'method' => $_SERVER['REQUEST_METHOD'],
+    'uri' => $_SERVER['REQUEST_URI']
+]);
+
+SimpleRouter::start();